refactor: separate DictyNews components into individual files for improved modularity and maintainability
